@@ -12,8 +12,8 @@ class VariantArray(GenomicArray):
     Required columns: chromosome, start, end, ref, alt
     """
 
-    _required_columns = ("chromosome", "start", "end", "ref", "alt")
-    _required_dtypes = (str, int, int, str, str)
+    _required_columns = ("chromosome", "start", "end", "ref", "alt")  # type: ignore
+    _required_dtypes = (str, int, int, str, str)  # type: ignore
     # Extra: somatic, zygosity, depth, alt_count, alt_freq
 
     def __init__(self, data_table, meta_dict=None):
